requirements.txt
pyjwt
cryptography
pynacl
httpx
websocket-client
dacite
